==> Evaluating_existing_models/Analytics_Scripts/1_Weartime_duration_overlap.ipynb <==
from google.colab import drive
drive.mount('/gdrive')

import os
from os import listdir, makedirs
from os.path import join, isfile, isdir
from datetime import datetime
import pandas as pd
from tqdm import tqdm

root_folder = 'drive/My Drive/LaTrobe/Projects/Accelerometer/OA activity data_La Trobe'
ActiGraph_raw = join(root_folder, 'ActiGraph_raw')
ActiGraph_60s_epoch = join(root_folder, 'ActiGraph_60s_epoch')
activPAL_raw = join(root_folder, 'activPAL_raw')
activPAL_15s_epoch = join(root_folder, 'activPAL_15s_epoch')

## ActiGraph_raw   

(HOA53) Start Date 9/03/2020  
(KOA37) Start Date 19/08/2019

listdir(ActiGraph_raw)

keys = [f.split(' ')[0] for f in listdir(ActiGraph_raw) if isfile(join(ActiGraph_raw, f))]
print(len(keys), len(listdir(ActiGraph_raw)))

print('KOA', len([a for a in keys if 'KOA' in a]))
print('HOA', len([a for a in keys if 'HOA' in a]))
print('OA', len([a for a in keys if 'HOA' not in a and 'KOA' not in a]))

Show captured dates

from datetime import datetime

def diff_dates(date1, date2):

  diff = date2 - date1

  days, seconds = diff.days, diff.seconds
  hours = days * 24 + seconds // 3600
  minutes = (seconds % 3600) // 60
  seconds = seconds % 60

  return '{}h {}m {}d'.format(hours, minutes, seconds)

# d1 = datetime(2013,1,1, 20, 3, 2)
# d2 = datetime(2013,1,13, 4, 2, 1)
# hours, minutes, seconds = diff_dates(d1, d2)
# print ('{}h {}m {}d days between {} and {}'.format(hours, minutes, seconds, d1, d2))
# print ("Happy programmer's day!")

def get_actigraph_duration(root_folder, files, date_list):

  for f in tqdm(files, desc=root_folder):

    # begin
    with open(join(root_folder, f)) as fle:
      for i, line in enumerate(fle):
        if i == 11:
          dt_begin = datetime.strptime(line.split(',')[0], '%d/%m/%Y %H:%M:%S.%f')
          dt_begin_formatted = dt_begin.strftime('%Y-%m-%d %H:%M:%S')
          # date_list.append([f.split()[0], dt_formatted])
          # dates[user_type].append(dt)
        elif i > 11:
          break

    # end
    with open(join(root_folder, f), 'rb') as flr:
      flr.seek(-2, os.SEEK_END)
      while flr.read(1) != b'\n':
          flr.seek(-2, os.SEEK_CUR)
      last_line = flr.readline().decode()

      dt_end = datetime.strptime(last_line.split(',')[0], '%d/%m/%Y %H:%M:%S.%f')
      dt_end_formatted = dt_end.strftime('%Y-%m-%d %H:%M:%S')
      # dates[user_type].append(dt)

    difference = diff_dates(dt_begin, dt_end)

    # Create data
    # print(f)
    date_list.append([f.split()[0], dt_begin_formatted, dt_end_formatted, difference])

  return date_list


# dates = {'HOA': [], 'KOA': [], 'OA': []}
date_list_global = []

# for HOA
user_type = 'HOA'
files = [f for f in listdir(ActiGraph_raw) if isfile(join(ActiGraph_raw, f)) and user_type in f]
date_list_global = get_actigraph_duration(ActiGraph_raw, files, date_list_global)

# for KOA
user_type = 'KOA'
files = [f for f in listdir(ActiGraph_raw) if isfile(join(ActiGraph_raw, f)) and user_type in f]
date_list_global = get_actigraph_duration(ActiGraph_raw, files, date_list_global)

# for OA
user_type = 'OA'
files = [f for f in listdir(ActiGraph_raw) if isfile(join(ActiGraph_raw, f)) and 'KOA' not in f and 'HOA' not in f]
date_list_global = get_actigraph_duration(ActiGraph_raw, files, date_list_global)


len(date_list_global)

actigraph_df = pd.DataFrame(date_list_global, columns=['Participant', 'actigraph_Start_time', 'actigraph_End_time', 'actigraph_Time_difference'])

## ActiGraph 60s

keys = [f.split(' ')[0] for f in listdir(ActiGraph_60s_epoch) if isfile(join(ActiGraph_60s_epoch, f))]
print(len(keys), len(listdir(ActiGraph_60s_epoch)))

print('KOA', len([a for a in keys if 'KOA' in a]))
print('HOA', len([a for a in keys if 'HOA' in a]))
print('OA', len([a for a in keys if 'HOA' not in a and 'KOA' not in a]))

## activPAL_raw  

Time conversion: https://stackoverflow.com/questions/1703505/excel-date-to-unix-timestamp  

  
Important:
OA 13 and 20 missing

listdir(activPAL_raw)

keys = [f.split('-')[0] for f in listdir(activPAL_raw) if isfile(join(activPAL_raw, f))]
print(len(keys), len(listdir(activPAL_raw)))

print('KOA', len([a for a in keys if 'KOA' in a]))
print('HOA', len([a for a in keys if 'HOA' in a]))
print('OA', len([a for a in keys if 'HOA' not in a and 'KOA' not in a]))

def get_activpal_duration(root_folder, files, date_list):

  for f in tqdm(files):

    # begin
    with open(join(root_folder, f)) as fle:
      for i, line in enumerate(fle):
        if i == 2:
          ts_begin = (float(line.split(';')[0].replace('\n', '')) - 25569) * 86400
          dt_begin = datetime.utcfromtimestamp(ts_begin).strftime('%Y-%m-%d %H:%M:%S')
          # date_actipal_list.append([f.split('-')[0], dt])
          # dates_actipal_raw[user_type].append(dt)
        elif i > 2:
          break

    # end
    with open(join(root_folder, f), 'rb') as flr:
      flr.seek(-2, os.SEEK_END)
      while flr.read(1) != b'\n':
          flr.seek(-2, os.SEEK_CUR)
      last_line = flr.readline().decode()

      ts_end = (float(last_line.split(';')[0].replace('\n', '')) - 25569) * 86400
      dt_end = datetime.utcfromtimestamp(ts_end).strftime('%Y-%m-%d %H:%M:%S')
      # dates[user_type].append(dt)

    difference = diff_dates(datetime.strptime(dt_begin, '%Y-%m-%d %H:%M:%S'), datetime.strptime(dt_end, '%Y-%m-%d %H:%M:%S'))

    # Create data
    # print(f)
    date_list.append([f.split('-')[0], dt_begin, dt_end, difference])

  return date_list


# dates_actipal_raw = {'HOA': [], 'KOA': [], 'OA': []}
date_actipal_list = []

# for HOA
user_type = 'HOA'
files = [f for f in listdir(activPAL_raw) if isfile(join(activPAL_raw, f)) and user_type in f]
date_actipal_list = get_activpal_duration(activPAL_raw, files, date_actipal_list)

# for KOA
user_type = 'KOA'
files = [f for f in listdir(activPAL_raw) if isfile(join(activPAL_raw, f)) and user_type in f]
date_actipal_list = get_activpal_duration(activPAL_raw, files, date_actipal_list)

# for OA
user_type = 'OA'
files = [f for f in listdir(activPAL_raw) if isfile(join(activPAL_raw, f)) and 'KOA' not in f and 'HOA' not in f]
date_actipal_list = get_activpal_duration(activPAL_raw, files, date_actipal_list)


activpal_df = pd.DataFrame(date_actipal_list, columns=['Participant', 'activpal_Start_time', 'activpal_End_time', 'activpal_Time_difference'])

activpal_df.head()

actigraph_df.head()

activpal_df.shape

actigraph_df.shape

Both_DFs = pd.merge(activpal_df,actigraph_df, how='outer',on=['Participant'])

Both_DFs.head(50)

Both_DFs.columns

Both_DFs['activpal_Start_time'] = pd.to_datetime(Both_DFs['activpal_Start_time'], format='%Y-%m-%d %H:%M:%S')
Both_DFs['activpal_End_time'] = pd.to_datetime(Both_DFs['activpal_End_time'], format='%Y-%m-%d %H:%M:%S')
Both_DFs['actigraph_Start_time'] = pd.to_datetime(Both_DFs['actigraph_Start_time'], format='%Y-%m-%d %H:%M:%S')
Both_DFs['actigraph_End_time'] = pd.to_datetime(Both_DFs['actigraph_End_time'], format='%Y-%m-%d %H:%M:%S')

## Overlap


from datetime import datetime
from collections import namedtuple
Range = namedtuple('Range', ['start', 'end'])

overlap_data = []

for index, row in Both_DFs.iterrows():

  if pd.isna(row['activpal_Start_time']) or pd.isna(row['activpal_End_time']) or pd.isna(row['actigraph_Start_time']) or pd.isna(row['actigraph_End_time']):
    overlap_data.append(None)
    continue

  r1 = Range(start=row['activpal_Start_time'], end=row['activpal_End_time'])
  r2 = Range(start=row['actigraph_Start_time'], end=row['actigraph_End_time'])

  latest_start = max(r1.start, r2.start)
  earliest_end = min(r1.end, r2.end)

  if latest_start > earliest_end:
    overlap_data.append(None)
    continue

  overlap = diff_dates(latest_start, earliest_end)
  overlap_data.append(overlap)

# Combine to dataframe
Both_DFs['Overlap'] = overlap_data

Both_DFs.to_csv('record_avaiability.csv')

## activPal 15s

keys = [f.split('-')[0] for f in listdir(activPAL_15s_epoch) if isfile(join(activPAL_15s_epoch, f))]
print(len(keys), len(listdir(activPAL_15s_epoch)))

keys

print('KOA', len([a for a in keys if 'KOA' in a]))
print('HOA', len([a for a in keys if 'HOA' in a]))
print('OA', len([a for a in keys if 'HOA' not in a and 'KOA' not in a]))

# KOA50 events
