==> Evaluating_existing_models/Analytics_Scripts/2_Combined_Sleep_time_generator.ipynb <==
# from google.colab import drive
# drive.mount('/gdrive')

import os
from os import listdir, makedirs
from os.path import join, isfile, isdir
from datetime import datetime
import pandas as pd
import numpy as np
from tqdm import tqdm

root_folder = 'drive/My Drive/LaTrobe/Projects/Accelerometer/OA activity data_La Trobe'
overlap_data_file = join(root_folder, 'Analytics_Scripts/metadata/record_avaiability.xlsx')
hkoa_data_file = join(root_folder, 'anthro.xlsx')
oa_data_folder = join(root_folder, 'sleep_oa')

main_df = pd.read_excel(overlap_data_file)

oa_data_files = { "OA{}".format(int(f.split('_')[-1].replace('.csv', ''))): f for f in listdir(oa_data_folder) if isfile(join(oa_data_folder, f))}

## Combine HOA, KOA and metadata

main_df.head()

hkoa_data_df.head()

main_df_merged = pd.merge(main_df, hkoa_data_df, left_on='Participant', right_on='pid', how='left')

main_df_merged.tail()

## Duplicate OA data with daily sleep times

oa_data_df = None

for key, fname in tqdm(oa_data_files.items()):

  temp_df = pd.read_csv(join(oa_data_folder, fname))
  temp_df['pid'] = key

  if oa_data_df is not None:
    oa_data_df = oa_data_df.append(temp_df)
  else:
    oa_data_df = temp_df

# Update datetime
oa_data_df['SLEEP_start'] = pd.to_datetime(oa_data_df['SLEEP_start'], format='%d%b%y:%H:%M:%S')
oa_data_df['SLEEP_finish'] = pd.to_datetime(oa_data_df['SLEEP_finish'], format='%d%b%y:%H:%M:%S')

oa_data_df = oa_data_df.rename(columns={'SLEEP_start': 'av_inbedtime', 'SLEEP_finish': 'av_outofbed_time'})

merged_df = pd.merge(main_df_merged, oa_data_df, left_on='pid', right_on='pid', how='left')

merged_df.head()

merged_df['av_inbedtime'] = merged_df['av_inbedtime_x'].astype(str) + merged_df['av_inbedtime_y'].astype(str)
merged_df['av_inbedtime'] = merged_df['av_inbedtime'].str.replace('nan', '')
merged_df['av_inbedtime'] = merged_df['av_inbedtime'].str.replace('NaT', '')

merged_df['av_outofbed_time'] = merged_df['av_outofbed_time_x'].astype(str) + merged_df['av_outofbed_time_y'].astype(str)
merged_df['av_outofbed_time'] = merged_df['av_outofbed_time'].str.replace('nan', '')
merged_df['av_outofbed_time'] = merged_df['av_outofbed_time'].str.replace('NaT', '')

del merged_df['av_inbedtime_x']
del merged_df['av_inbedtime_y']
del merged_df['av_outofbed_time_x']
del merged_df['av_outofbed_time_y']
del merged_df['pid']

merged_df.tail(40)

merged_df.to_excel('sleep_time.xlsx', index=None)
