==> Evaluating_existing_models/Analytics_Scripts/3_Generate_ref_for_actigraph.ipynb <==
from google.colab import drive
drive.mount('/gdrive')

import os
from os import listdir, makedirs
from os.path import join, isfile, isdir
from datetime import datetime
import pandas as pd
import numpy as np
from tqdm import tqdm

# Input
root_folder = '/content/drive/My Drive/OA_Gihan/OA_all/OA activity data_La Trobe'
ActiGraph_60s_epoch = join(root_folder, 'ActiGraph_60s_epoch')
# activPAL_15s_epoch = join(root_folder, 'activPAL_15s_epoch')

# sleep time input_data
sleep_time_file = join(root_folder, 'Analytics_Scripts/Rashmika_Analysis/metadata/Copy of sleep_time.xlsx')

# Output
Actigraph_output_folder = join(root_folder, 'Analytics_Scripts/Rashmika_Analysis/processed_data/actigraph/epoch_data/60s_ref')

from google.colab import drive
drive.mount('/content/drive')

### Load Sleep Data

sleep_df = pd.read_excel(sleep_time_file)

sleep_df.head()

### MET Equation

def get_freedson_vm3_combination_11_energy_expenditure(row):

    if hip_epoch_data['waist_vm_cpm'][row.name] <= 2453:
        # METs = 0.001092(VA) + 1.336129  [capped at 2.9999, where VM3 < 2453]
        met_value = (0.001092 * hip_epoch_data['waist_cpm'][row.name]) + 1.336129
        met_value = met_value if met_value < 2.9999 else 2.9999
    else:
        # METs = 0.000863(VM3) + 0.668876 [where VM3 ≥ 2453]
        met_value = 0.000863 * hip_epoch_data['waist_vm_60'][row.name] + 0.668876

    return met_value

### Process ref values

epoch_files = [f for f in listdir(ActiGraph_60s_epoch) if isfile(join(ActiGraph_60s_epoch, f))]
print(epoch_files)

epoch_files[156]

#check#OA46
# oafile = join(ActiGraph_60s_epoch, epoch_files[156])
# print(oafile)
# oa46 = pd.read_csv(oafile, skiprows=11)
# oa46.head()

sleep_df.loc[sleep_df['Participant'] == 'OA46']

for f in tqdm(epoch_files):

  user_id = f.split()[0]

  hip_epoch_data = pd.read_csv(join(ActiGraph_60s_epoch, f), skiprows=11, usecols=[0, 1, 2], header=None)
  # Axis 1 (y) - Goes through head and middle of feet
  # Axis 2 (x) - Goes through 2 hips
  # Axis 3 (z) - Goes through front and back of the stomach
  hip_epoch_data.columns = ['AxisY', 'AxisX', 'AxisZ']

  # Create time column
  print(user_id)
  if user_id != 'OA46_fixed60sec.csv':
    start_time = sleep_df.loc[sleep_df['Participant'] == user_id].iloc[0]['actigraph_Start_time']
  else:
    start_time = "07-03-2016  4:00:00"

  hip_epoch_data['Time_stamp'] = list(pd.date_range(start_time, periods=hip_epoch_data.shape[0], freq='min'))
  # print(hip_epoch_data['Time_stamp'])
  # break
  hip_epoch_data['waist_vm_60'] = np.sqrt([(hip_epoch_data.AxisX ** 2) + (hip_epoch_data.AxisY ** 2) + (hip_epoch_data.AxisZ ** 2)])[0]
  hip_epoch_data['waist_vm_cpm'] = hip_epoch_data['waist_vm_60']  # cpm = counts per minute
  hip_epoch_data['waist_cpm'] = hip_epoch_data.AxisY

  hip_epoch_data['waist_ee'] = hip_epoch_data.apply(get_freedson_vm3_combination_11_energy_expenditure, axis=1)

  # # Use following snippet if need the reference data for granular time_epochs
  # multiplication_factor = int(60 / time_epoch)
  # epoch_hip_epoch_data = epoch_hip_epoch_data.reindex(np.repeat(epoch_hip_epoch_data.index.values, multiplication_factor), method='ffill')

  # save output file
  # Input example: 'HOA1 (2019-05-06)60sec.csv'
  output_filename = f.replace('.csv', '_60sREF.csv')
  hip_epoch_data.to_csv(join(Actigraph_output_folder, output_filename), sep=',', index=None)
  del hip_epoch_data
