==> Evaluating_existing_models/Analytics_Scripts/3_Generate_ref_for_actipal.ipynb <==
from google.colab import drive
drive.mount('/gdrive')

import os
from os import listdir, makedirs
from os.path import join, isfile, isdir
from datetime import datetime
import pandas as pd
import numpy as np
from tqdm import tqdm

# Input
root_folder = '/content/drive/My Drive/OA_Gihan/OA_all/OA activity data_La Trobe'
# ActiGraph_60s_epoch = join(root_folder, 'ActiGraph_60s_epoch')
activPAL_15s_epoch = join(root_folder, 'activPAL_15s_epoch')

# sleep time input_data
sleep_time_file = '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Rashmika_analysis/metadata/sleep_time.xlsx'

# Output
ActiPAL_output_folder = join(root_folder, '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/Data/processed_data/ActiPAL/epoch_data/15s_ref')

from google.colab import drive
drive.mount('/content/drive')

### Load Sleep Data

sleep_df = pd.read_excel(sleep_time_file)

sleep_df.head()

### MET from Activity Score



def get_met_from_activity_counts_1min(score_df,step_count_df):
  output_df = (score_df-(4*step_count_df*4*15/120))/(15*(1-(step_count_df*4/120)))
  return output_df

def get_met_from_activity_counts_15s(score_df,step_count_df):
  output_df = (score_df-(4*step_count_df*15/120))/(15*(1-(step_count_df/120)))
  return output_df

def get_met_from_first_part(score_df):
  output_df = score_df/15
  return output_df

def get_met_from_step_count(row):
  if (thigh_epoch_data['Step Count'][row.name]==0):
    met_value = 1.25
  else:
    met_value = ((1.4*0.0167)+(4-1.4)*(thigh_epoch_data['Step Count'][row.name]*4/120)*0.0167)*60
  return met_value

input_data = pd.DataFrame()
input_data['score'] = [22.2,22.73,18.75,18.75,18.75,18.75,18.75,18.75]
input_data['step'] = [2,2,0,0,0,0,0,0]

print(get_met_from_first_part(input_data['score']))
print(get_met_from_activity_counts_15s(input_data['score'],input_data['step']))
print(get_met_from_activity_counts_1min(input_data['score'],input_data['step']))

### Process ref values

epoch_files = [f for f in listdir(activPAL_15s_epoch) if (isfile(join(activPAL_15s_epoch, f)) and ('Events' not in f))]

print(epoch_files)

epoch_files[156]

#check#OA46
# oafile = join(ActiGraph_60s_epoch, epoch_files[156])
# print(oafile)
# oa46 = pd.read_csv(oafile, skiprows=11)
# oa46.head()

sleep_df.loc[sleep_df['Participant'] == 'OA46']

for f in tqdm(epoch_files):

  user_id = f.split('-')[0]

  thigh_epoch_data = pd.read_csv(join(activPAL_15s_epoch, f), skiprows=11, usecols=[1,2], header=None)
  thigh_epoch_data.columns = ['Step Count', 'Activity_score']

  # Create time column
  print(user_id)
  if(user_id=='OA13' or user_id=='OA20'):continue
  start_time = sleep_df.loc[sleep_df['Participant'] == user_id].iloc[0]['activpal_Start_time']
  # if user_id != 'OA46_fixed60sec.csv':
  #   start_time = sleep_df.loc[sleep_df['Participant'] == user_id].iloc[0]['actigraph_Start_time']
  # else:
  #   start_time = "07-03-2016  4:00:00"

  thigh_epoch_data['Time_stamp'] = list(pd.date_range(start_time, periods=thigh_epoch_data.shape[0], freq='15S'))
  # print(hip_epoch_data['Time_stamp'])
  # break

  # thigh_epoch_data['thigh_EE'] = get_met_from_activity_counts_15s(thigh_epoch_data['Activity_score'],thigh_epoch_data['Step Count'])
  thigh_epoch_data['thigh_EE'] = thigh_epoch_data.apply(get_met_from_step_count,axis=1)
  # print(thigh_epoch_data)
  # break
  # hip_epoch_data['waist_vm_60'] = np.sqrt([(hip_epoch_data.AxisX ** 2) + (hip_epoch_data.AxisY ** 2) + (hip_epoch_data.AxisZ ** 2)])[0]
  # hip_epoch_data['waist_vm_cpm'] = hip_epoch_data['waist_vm_60']  # cpm = counts per minute
  # hip_epoch_data['waist_cpm'] = hip_epoch_data.AxisY

  # hip_epoch_data['waist_ee'] = hip_epoch_data.apply(get_freedson_vm3_combination_11_energy_expenditure, axis=1)

  # # Use following snippet if need the reference data for granular time_epochs
  # multiplication_factor = int(60 / time_epoch)
  # epoch_hip_epoch_data = epoch_hip_epoch_data.reindex(np.repeat(epoch_hip_epoch_data.index.values, multiplication_factor), method='ffill')

  # save output file
  # Input example: 'HOA1 (2019-05-06)60sec.csv'
  output_filename = f.replace('.csv', '_15sREF.csv')
  thigh_epoch_data.to_csv(join(ActiPAL_output_folder, output_filename), sep=',', index=None)
  del thigh_epoch_data
