==> Evaluating_existing_models/Analytics_Scripts/6_Analyze_combined_data.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import os
from os import listdir, makedirs
from os.path import join, isfile, isdir
from datetime import datetime
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from tqdm import tqdm

# Input
root_folder = 'drive/My Drive/LaTrobe/Projects/Accelerometer/OA activity data_La Trobe'
Actigraph_combine_data_folder = join(root_folder, 'Analytics_Scripts/Rashmika_analysis/processed_data/actigraph/combined_data/Epoch30')

### Analyze_Intensity_over_time

files = [f for f in listdir(Actigraph_combine_data_folder)]
len(files)

df = pd.read_csv(join(Actigraph_combine_data_folder, files[4]))
print(df.columns)
corr = df.corr()
corr.to_csv('test.csv')
print(corr.head())

for f in files:

  df = pd.read_csv(join(Actigraph_combine_data_folder, f))
  print(df.columns)
  corr = df[['waist_ee', 'enmo', 'raw_XMean', 'raw_YMean', 'raw_ZMean',
       'raw_XVar', 'raw_YVar', 'raw_ZVar', 'raw_X10perc', 'raw_X25perc',
       'raw_X50perc', 'raw_X75perc', 'raw_X90perc', 'raw_Y10perc',
       'raw_Y25perc', 'raw_Y50perc', 'raw_Y75perc', 'raw_Y90perc',
       'raw_Z10perc', 'raw_Z25perc', 'raw_Z50perc', 'raw_Z75perc',
       'raw_Z90perc', 'raw_X_cov', 'raw_Y_cov', 'raw_Z_cov']].corr()
  corr.to_csv('test.csv')
  print(corr.head())

  break

for _, row in tqdm(sleep_df.iterrows(), total=sleep_df.shape[0]):

  user_id = row['Participant']
  begin_time = row['begin_time']
  end_time = row['end_time']

  # get respective epoch dataset
  # there is a problem with key OA46
  if user_id not in data_dict.keys():
    continue

  epoch_df = pd.read_csv(data_dict[user_id]['epoch'])
  epoch_df = epoch_df[(begin_time < epoch_df['Time_stamp']) & (epoch_df['Time_stamp'] < end_time)]

  plt.figure()
  epoch_df['waist_ee'].plot()
  plt.title('{}: {} to {}'.format(user_id, begin_time, end_time))
  plt.savefig(join(Actigraph_image_output, '{}: {} to {}.jpeg'.format(user_id, begin_time, end_time).replace(':', '_').replace(' ', '-')))

  plt.clf()
  plt.close()
