==> Evaluating_existing_models/Analytics_Scripts/7_parameters for hibbing 1s.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import os
from os import listdir, makedirs
from os.path import join, isfile, isdir, exists
from datetime import datetime
import pandas as pd
import numpy as np
import math
from tqdm import tqdm

# Input
root_folder = '/content/drive/My Drive/OA activity data_La Trobe'
ActiGraph_60s_epoch_proc_folder = '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Rashmika_analysis/processed_data/actigraph/epoch_data/60s_ref'
ActiGraph_60hz_raw_folder = '/content/drive/My Drive/left_graph_raw'

# sleep time input_data
sleep_time_file = '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Rashmika_analysis/metadata/record_availability_final.csv'

# Output
Actigraph_output_folder = '/content/drive/My Drive/processed data/hibbing/'

### Load Sleep Data

raw_files = [f for f in listdir(ActiGraph_60hz_raw_folder) if isfile(join(ActiGraph_60hz_raw_folder, f))]
epoch_files = [f for f in listdir(ActiGraph_60s_epoch_proc_folder) if isfile(join(ActiGraph_60s_epoch_proc_folder, f))]

data_dict = {f.split()[0]: {'raw': join(ActiGraph_60hz_raw_folder, f)} for f in raw_files}
for ef in epoch_files:
  key = ef.split()[0]
  if key in data_dict.keys():
    data_dict[key]['epoch'] = join(ActiGraph_60s_epoch_proc_folder, ef)

print(len(raw_files))
print(len(epoch_files))
print(data_dict)

sleep_df = pd.read_csv(sleep_time_file)
sleep_df.head(2)
print(len(sleep_df))

### Generate statistical features

FREQUENCY = 60 #HZ amount of the accelerometer
REFERENCE_EPOCH = 60 #Epoch value that EE is given
TIME_EPOCH_DICT = {
    'Epoch1': FREQUENCY * 1,
    # 'Epoch5': FREQUENCY * 5,
    # 'Epoch30': FREQUENCY * 30,
    # 'Epoch60': FREQUENCY * 60
}

def hibbing_ee_prediction_hip_1(row):
 
  if (epoch_df_current['CWR_threshold_enmo'][row.name]<=15.8):
    predicted_ee = 1.48 +0.01405*(epoch_df_current['enmo'][row.name])
  else:
    predicted_ee = 1.76 +0.04844*(epoch_df_current['enmo'][row.name])-0.00015*(epoch_df_current['enmo'][row.name]*epoch_df_current['enmo'][row.name]) + 0.0000003*(epoch_df_current['enmo'][row.name]*epoch_df_current['enmo'][row.name]*epoch_df_current['enmo'][row.name])

  return predicted_ee



# completed_folder = Actigraph_output_folder+'/Epoch1/'
completed_folder = '/content/drive/My Drive/processed data/hibbing/Epoch1'
completed_files = [f for f in listdir(completed_folder) if isfile(join(completed_folder, f))]
completed_ids = [f.split('_')[0] for f in listdir(completed_folder) if isfile(join(completed_folder, f))]
completed_ids = list(set(completed_ids))
print(len(completed_files))
completed_ids.sort()
print(completed_ids)
print(len(completed_ids))

for _, row in tqdm(sleep_df.iterrows(), total=sleep_df.shape[0]):
  user_id = row['Participant']
  print(_,user_id)
  if(user_id in completed_ids):
    continue
  # if(_<1262):continue
  # if(user_id != 'OA35'):continue
  try:
  
    begin_time = row['begin_time']
    end_time = row['end_time']

    # if('HOA1' != user_id):
    #   continue
    print(user_id)

    # get respective epoch dataset
    epoch_df = pd.read_csv(data_dict[user_id]['epoch'])
    epoch_df = epoch_df[(begin_time < epoch_df['Time_stamp']) & (epoch_df['Time_stamp'] < end_time)]
    # print(list(epoch_df.index)[0])
    # continue
    # get respective raw dataset
    raw_begin_index = list(epoch_df.index)[0] * REFERENCE_EPOCH * FREQUENCY
    raw_data_length = epoch_df.shape[0] * REFERENCE_EPOCH * FREQUENCY
    raw_df_orig = pd.read_csv(data_dict[user_id]['raw'], skiprows=11+raw_begin_index, nrows=raw_data_length, header=None)
    raw_df_orig.columns = ['Timestamp', 'X', 'Y', 'Z']
    # print(raw_df_orig.head(20))
    

    """
    Process RAW data as required by the model features
    """

    # VM and MANGLE needs for calculations by Staudenmayer model
    # Calculate the vector magnitude from X, Y, Z raw readings
    # Calculate the angle of arcsin from X and VM, arcsin(axis used/vector magnitude)/(pi/2)
    raw_df_orig['vm'] = np.sqrt([(raw_df_orig.X ** 2) + (raw_df_orig.Y ** 2) + (raw_df_orig.Z ** 2)])[0]
    # raw_df_orig['angle'] = (90 * np.arcsin(raw_df_orig.X / raw_df_orig['vm'])) / (math.pi / 2)
    # print(raw_df_orig['vm'].head(20))
    # Calculate ENMO (Euclidean Norm Minus One) reguired for Hilderbrand model
    raw_df_orig['enmo'] = np.sqrt([(raw_df_orig.X ** 2) + (raw_df_orig.Y ** 2) + (raw_df_orig.Z ** 2)])[0] - 1
    raw_df_orig.loc[raw_df_orig.enmo < 0, 'enmo'] = 0
    # print(raw_df_orig['enmo'].head(20))

    # For each epoch given
    for epoch_name, epoch_duration in TIME_EPOCH_DICT.items():
      print(epoch_name,epoch_duration)
      
      # continue
      # Create output path for the combined file
      output_file_name = join(Actigraph_output_folder, epoch_name,'{}_{}_to_{}_{}.csv'.format(user_id, begin_time, end_time, epoch_name).replace(':', '-').replace(' ', '_'))
      # if(output_file_name in completed_files):
      #   print('skipping')
      #   continue
      if not exists(join(Actigraph_output_folder, epoch_name)):
        makedirs(join(Actigraph_output_folder, epoch_name))


      # Copy raw data
      raw_df = raw_df_orig.copy()
      # print(raw_df.head(10))
      # print(raw_df.index)
      raw_df.reset_index(drop=True, inplace=True)
      # print(raw_df.head(10))
      # print(raw_df.index)
      
      """
      Calculate the statistical inputs (Features)
      """
      # print(raw_df.index // epoch_duration)
      
      hip_grouped_temp = raw_df.groupby(raw_df.index // epoch_duration)
      # k=0
      # for name,group in hip_grouped_temp:
      #   print(name)
      #   print(group)
      #   k=k+1
      #   if(k==100):break
      # break
      aggregated_hip = pd.DataFrame()

      # General stats
      aggregated_hip['mvm'] = hip_grouped_temp['vm'].mean()
      aggregated_hip['enmo'] = hip_grouped_temp['enmo'].mean()
      aggregated_hip['enmo'] = aggregated_hip['enmo'].fillna(1)

    

      """
      Repeat rows of epoch file as required to supplement raw data.
      """
      # Number of rows for each epoch
      epoch_df_current = epoch_df.copy()
      # print(epoch_df_current.head())
      epoch_df_current.reset_index(drop=True, inplace=True)
      # print(epoch_df_current.head())
      
      multiplication_factor = int(REFERENCE_EPOCH / (epoch_duration/FREQUENCY))#ex: 60/(1800/60) for 30 epochs, ultimately 60/30
      # print('mul',[multiplication_factor,REFERENCE_EPOCH,epoch_duration,FREQUENCY])
      # print('before',epoch_df_current.head())
      epoch_df_current = epoch_df_current.reindex(np.repeat(epoch_df_current.index.values, multiplication_factor), method='ffill')
      # print('after',epoch_df_current.head())
      epoch_df_current['waist_ee_final'] = epoch_df_current['waist_ee']/multiplication_factor
      # print('EE',epoch_df_current['waist_ee'])

      epoch_df_current.reset_index(drop=True, inplace=True)
      


      epoch_df_current['enmo'] = aggregated_hip['enmo']
      epoch_df_current['mvm'] = aggregated_hip['mvm']
      

      CWR_list_enmo = []
      CWR_list_gvm = []
      for i in range(len(epoch_df_current)):

        mean_w_e = epoch_df_current['enmo'].iloc[i:i+10].mean()
        std_w_e = epoch_df_current['enmo'].iloc[i:i+10].std()
        if(mean_w_e==0):
          cov_w_e = 0
        else:
          cov_w_e = (std_w_e/mean_w_e)*100
        
        
        if(math.isnan(cov_w_e)):cov_w_e = 0
        CWR_list_enmo.append(cov_w_e)

        mean_w_g = epoch_df_current['mvm'].iloc[i:i+10].mean()
        std_w_g = epoch_df_current['mvm'].iloc[i:i+10].std()
        if(mean_w_g==0):
          cov_w_g = 0
        else:
          cov_w_g = (std_w_g/mean_w_g)*100
        
        
        if(math.isnan(cov_w_g)):cov_w_g = 0
        CWR_list_gvm.append(cov_w_g)
      print('gvm')
      print(len(CWR_list_gvm),len(epoch_df_current))
      print('enmo')
      print(len(CWR_list_enmo),len(epoch_df_current))
      # print(epoch_df_current.head())

      epoch_df_current['CWR_threshold_enmo'] = CWR_list_enmo
      epoch_df_current['CWR_threshold_gvm'] = CWR_list_gvm


      # epoch_df_current['EE_hibbing'] = hibbing_ee_prediction_hip_1(epoch_df_current['enmo'],epoch_df_current['mvm'],['CWR_threshold_enmo'],['CWR_threshold_gvm'])
      epoch_df_current['EE_hibbing'] = epoch_df_current.apply(hibbing_ee_prediction_hip_1,axis=1)








      # print('epoch_df', epoch_df.shape)
      # print('epoch_df_current', epoch_df_current.shape, multiplication_factor)
      # print('raw', aggregated_hip.shape)
      # print('ag_hip', aggregated_hip.head())
      # print('after',epoch_df_current.head())

      epoch_df_current.to_csv(output_file_name, index=None)

  except SyntaxError as e:
    print('Error')
    f = open('errors.txt','w+')
    f.write(str(user_id)+" "+str(e)+" \n")
    f.close()


f_c = '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/hibbing/Epoch1/HOA1_2019-05-06_15-00-00_to_2019-05-06_22-00-00_Epoch1.csv'
f_c_df = pd.read_csv(f_c)
f_c_df.head()


test = [0.051245,0.046331,0.078401,0.051831,0.067690]
t_df = pd.DataFrame(test)
s = t_df.rolling(2)
print(s.head())

CWR_list_enmo = []
CWR_list_gvm = []
for i in range(len(f_c_df)):

  mean_w_e = f_c_df['enmo'].iloc[i:i+10].mean()
  std_w_e = f_c_df['enmo'].iloc[i:i+10].std()
  if(mean_w_e==0):
    cov_w_e = 0
  else:
    cov_w_e = (std_w_e/mean_w_e)*100
  
  
  if(math.isnan(cov_w_e)):cov_w_e = 0
  CWR_list_enmo.append(cov_w_e)

  mean_w_g = f_c_df['mvm'].iloc[i:i+10].mean()
  std_w_g = f_c_df['mvm'].iloc[i:i+10].std()
  if(mean_w_g==0):
    cov_w_g = 0
  else:
    cov_w_g = (std_w_g/mean_w_g)*100
  
  
  if(math.isnan(cov_w_g)):cov_w_g = 0
  CWR_list_gvm.append(cov_w_g)
 
# print(CWR_list_enmo[-1])

# print(CWR_list_enmo[-1])
# print(len(CWR_enmo_df),len(f_c_df))


# print(CWR_list_gvm[-1])

# print(CWR_list_gvm[-1])
# print(len(CWR_gvm_df),len(f_c_df))

for i,row in f_c_df.iterrows():
  next_10_1s_enmo = []
  next_10_1s_gvm = []
  # print(row['enmo'])
  for k in range(i,i+10):
    try:
        row['enmo'].sh
        # next_10_1s_enmo.append(f_c_df['enmo'].iloc[[k]])
    except IndexError:
   
        # next_10_1s_enmo.append(0)
  enmo_df = pd.DataFrame(next_10_1s_enmo)
  print(i,next_10_1s_enmo)
  print('cv',enmo_df.cov())
  if(i==10):break
      # f.write(str(row))
      # next_itms = []
      # for k in range(i,i+10):
      #   try:
      #     # next_itms+=" "+str(epoch_df_current.iloc[[k]]['mvm'])
      #     next_itms.append(epoch_df_current['mvm'].iloc[[k]])
      #   except IndexError:
      #     next_itms+="0"
      #     next_itms.append(0)
      #   print('next',epoch_df_current.iloc[[k]]['mvm'])
      # row['next_items'] = next_itms
      # f.write(str(i)+" "+str(next_itms)+" "+str(len(next_itms)))
      # if(i==20):break
      # print('nxt',next_itms)


f = open('/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/hibbing/file.txt','w')
f.write("hi")
f.close()

s = [1,2,3,2,2]
print("ss"+str(s))

df1 = pd.DataFrame()
x = np.linspace(0,1000,100)
df1['x'] = x

y = np.linspace(0,500,100)
df1['y'] = y

z = np.linspace(0,3000,100)
df1['z'] = z

gdf = df1.groupby(df1.index//10)

for name,group in gdf:
  print(name)
  print(group)

print(gdf.groups)
