==> Evaluating_existing_models/Analytics_Scripts/8_inputs_for_ANN_hip_2016.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import os
from os import listdir, makedirs
from os.path import join, isfile, isdir, exists
from datetime import datetime
import pandas as pd
import numpy as np
import math
from tqdm import tqdm

# Input
root_folder = '/content/drive/My Drive/OA_Gihan/OA_all/OA activity data_La Trobe'
ActiGraph_30s_hip_parameters = '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/caron_and_montoye/Epoch30'

# Output
ActiGraph_output_folder = '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/caron_and_montoye/inputs_for_ANN_hip_2016'



# file = '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/caron_and_montoye/Epoch30/HOA1_2019-05-06_15-00-00_to_2019-05-06_22-00-00_Epoch30.csv'
# fe_data = pd.read_csv(file)
# fe_data.head(10)
# fe_data['raw_X10perc']

parameter_files = [f for f in listdir(ActiGraph_30s_hip_parameters) if isfile(join(ActiGraph_30s_hip_parameters, f))]
print(parameter_files)

for f in tqdm(parameter_files):
  fe_data = pd.read_csv(join(ActiGraph_30s_hip_parameters, f))
  # print(fe_data.head())


  columns = ['TenPct_AGhipX', 
             'TenPct_AGhipY', 
             'TenPct_AGhipZ', 
             'TwentyFivePct_AGhipX', 
             'TwentyFivePct_AGhipY', 
             'TwentyFivePct_AGhipZ',
             'FiftyPct_AGhipX', 
             'FiftyPct_AGhipY', 
             'FiftyPct_AGhipZ', 
             'SeventyFivePct_AGhipX', 
             'SeventyFivePct_AGhipY', 
             'SeventyFivePct_AGhipZ', 
             'NinetyPct_AGhipX', 
             'NinetyPct_AGhipY', 
             'NinetyPct_AGhipZ']

  result_data = pd.DataFrame(columns=columns)
  result_data['TenPct_AGhipX'] = fe_data['raw_X10perc']
  result_data['TenPct_AGhipY'] = fe_data['raw_Y10perc']
  result_data['TenPct_AGhipZ'] = fe_data['raw_Z10perc']
  result_data['TwentyFivePct_AGhipX'] = fe_data['raw_X25perc']
  result_data['TwentyFivePct_AGhipY'] = fe_data['raw_Y25perc']
  result_data['TwentyFivePct_AGhipZ'] = fe_data['raw_Z25perc']
  result_data['FiftyPct_AGhipX'] = fe_data['raw_X50perc']
  result_data['FiftyPct_AGhipY'] = fe_data['raw_Y50perc']
  result_data['FiftyPct_AGhipZ'] = fe_data['raw_Z50perc']
  result_data['SeventyFivePct_AGhipX'] = fe_data['raw_X75perc']
  result_data['SeventyFivePct_AGhipY'] = fe_data['raw_Y75perc']
  result_data['SeventyFivePct_AGhipZ'] = fe_data['raw_Z75perc']
  result_data['NinetyPct_AGhipX'] = fe_data['raw_X90perc']
  result_data['NinetyPct_AGhipY'] = fe_data['raw_Y90perc']
  result_data['NinetyPct_AGhipZ'] = fe_data['raw_Z90perc']

  output_f = f.replace('.csv', '_ANN_input_hip_2016.txt')
  output_filename = join(ActiGraph_output_folder, output_f)

  result_data.to_csv(output_filename, sep='\t', index=False)
