==> Evaluating_existing_models/Analytics_Scripts/8_inputs_for_ANN_thigh_2015_feature_set_3.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import os
from os import listdir, makedirs
from os.path import join, isfile, isdir, exists
from datetime import datetime
import pandas as pd
import numpy as np
import math
from tqdm import tqdm

# Input
root_folder = '/content/drive/My Drive/OA_Gihan/OA_all/OA activity data_La Trobe'
ActiPAL_30s_thigh_parameters = '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/montoye_thigh/Epoch30'

# Output
ActiPAL_output_folder = '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/montoye_thigh/inputs_for_ANN_thigh_2015_feature_set_3'



# file = '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/caron_and_montoye/Epoch30/HOA1_2019-05-06_15-00-00_to_2019-05-06_22-00-00_Epoch30.csv'
# fe_data = pd.read_csv(file)
# fe_data.head(10)
# fe_data['raw_X10perc']

parameter_files = [f for f in listdir(ActiPAL_30s_thigh_parameters) if isfile(join(ActiPAL_30s_thigh_parameters, f))]
print(len(parameter_files))

completed_files = [f for f in listdir(ActiPAL_output_folder) if isfile(join(ActiPAL_output_folder, f))]
print(len(completed_files))

A = [1]*10
print(A)

for f in tqdm(parameter_files):
  try:
    fe_data = pd.read_csv(join(ActiPAL_30s_thigh_parameters, f))
    # print(fe_data.head())
    columns = ['Participant', 
              'Activity',
            'OC_METs', 
            'TimeSec', 
            'Mean_AGthighX', 
            'Mean_AGthighY', 
            'Mean_AGthighZ', 
            'Var_AGthighX', 
            'Var_AGthighY', 
            'Var_AGthighZ',
              'Max_AGthighX',
              'Max_AGthighY',	
              'Max_AGthighZ',
              'Min_AGthighX',
              'Min_AGthighY',	
              'Min_AGthighZ']
    result_data = pd.DataFrame(columns=columns)
    result_data['Participant'] = [1]*len(fe_data)
    result_data['Activity'] = ['test_activity']*len(fe_data)
    result_data['OC_METs'] = [10]*len(fe_data)
    result_data['TimeSec'] = [100]*len(fe_data)
    result_data['Mean_AGthighX'] = fe_data['raw_XMean']
    result_data['Mean_AGthighY'] = fe_data['raw_YMean']
    result_data['Mean_AGthighZ'] = fe_data['raw_ZMean']
    result_data['Var_AGthighX'] = fe_data['raw_XVar']
    result_data['Var_AGthighY'] = fe_data['raw_YVar']
    result_data['Var_AGthighZ'] = fe_data['raw_ZVar']
    result_data['Max_AGthighX'] = fe_data['raw_Xmax']
    result_data['Max_AGthighY'] = fe_data['raw_Ymax']
    result_data['Max_AGthighZ'] = fe_data['raw_Zmax']
    result_data['Min_AGthighX'] = fe_data['raw_Xmin']
    result_data['Min_AGthighY'] = fe_data['raw_Ymin']
    result_data['Min_AGthighZ'] = fe_data['raw_Zmin']

    # print(result_data.head())
    

    output_f = f.replace('.csv', '_ANN_input_thigh_2015_feature_set_3.txt')
    output_filename = join(ActiPAL_output_folder, output_f)

    result_data.to_csv(output_filename, sep='\t', index=False, line_terminator='\n')
  except Exception as e:
    f = open('errors.txt','w+')
    f.write(" "+str(e)+" \n")
    f.close()
