==> Evaluating_existing_models/Analytics_Scripts/8_inputs_for_ANN_thigh_2016.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import os
from os import listdir, makedirs
from os.path import join, isfile, isdir, exists
from datetime import datetime
import pandas as pd
import numpy as np
import math
from tqdm import tqdm

# Input
root_folder = '/content/drive/My Drive/OA_Gihan/OA_all/OA activity data_La Trobe'
ActiPAL_30s_thigh_parameters = '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/montoye_thigh/Epoch30'

# Output
ActiPAL_output_folder = '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/montoye_thigh/inputs_for_ANN_thigh_2016'



# file = '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/caron_and_montoye/Epoch30/HOA1_2019-05-06_15-00-00_to_2019-05-06_22-00-00_Epoch30.csv'
# fe_data = pd.read_csv(file)
# fe_data.head(10)
# fe_data['raw_X10perc']

parameter_files = [f for f in listdir(ActiPAL_30s_thigh_parameters) if isfile(join(ActiPAL_30s_thigh_parameters, f))]
print(parameter_files)

for f in tqdm(parameter_files):
  fe_data = pd.read_csv(join(ActiPAL_30s_thigh_parameters, f))
  # print(fe_data.head())
  columns = ['TenPct_AGthighX', 
           'TenPct_AGthighY', 
           'TenPct_AGthighZ', 
           'TwentyFivePct_AGthighX', 
           'TwentyFivePct_AGthighY', 
           'TwentyFivePct_AGthighZ', 
           'FiftyPct_AGthighX', 
           'FiftyPct_AGthighY', 
           'FiftyPct_AGthighZ', 
           'SeventyFivePct_AGthighX', 
           'SeventyFivePct_AGthighY', 
           'SeventyFivePct_AGthighZ', 
           'NinetyPct_AGthighX', 
           'NinetyPct_AGthighY', 
           'NinetyPct_AGthighZ']

  result_data = pd.DataFrame(columns=columns)
  result_data['TenPct_AGthighX'] = fe_data['raw_X10perc']
  result_data['TenPct_AGthighY'] = fe_data['raw_Y10perc']
  result_data['TenPct_AGthighZ'] = fe_data['raw_Z10perc']
  result_data['TwentyFivePct_AGthighX'] = fe_data['raw_X25perc']
  result_data['TwentyFivePct_AGthighY'] = fe_data['raw_Y25perc']
  result_data['TwentyFivePct_AGthighZ'] = fe_data['raw_Z25perc']
  result_data['FiftyPct_AGthighX'] = fe_data['raw_X50perc']
  result_data['FiftyPct_AGthighY'] = fe_data['raw_Y50perc']
  result_data['FiftyPct_AGthighZ'] = fe_data['raw_Z50perc']
  result_data['SeventyFivePct_AGthighX'] = fe_data['raw_X75perc']
  result_data['SeventyFivePct_AGthighY'] = fe_data['raw_Y75perc']
  result_data['SeventyFivePct_AGthighZ'] = fe_data['raw_Z75perc']
  result_data['NinetyPct_AGthighX'] = fe_data['raw_X90perc']
  result_data['NinetyPct_AGthighY'] = fe_data['raw_Y90perc']
  result_data['NinetyPct_AGthighZ'] = fe_data['raw_Z90perc']

  output_f = f.replace('.csv', '_ANN_input_thigh_2016.txt')
  output_filename = join(ActiPAL_output_folder, output_f)

  result_data.to_csv(output_filename, sep='\t', index=False)


for f in tqdm(parameter_files):
  fe_data = pd.read_csv(join(ActiPAL_30s_thigh_parameters, f))
  # print(fe_data.head())

  columns = ['TenPct_GEleftX', 'TenPct_GEleftY', 'TenPct_GEleftZ', 'TwentyFivePct_GEleftY',
                   'TwentyFivePct_GEleftZ',
                   'FiftyPct_GEleftX', 'FiftyPct_GEleftY', 'FiftyPct_GEleftZ', 'SeventyFivePct_GEleftX',
                   'SeventyFivePct_GEleftY',
                   'SeventyFivePct_GEleftZ', 'NinetyPct_GEleftX', 'NinetyPct_GEleftY', 'NinetyPct_GEleftZ']

  result_data = pd.DataFrame(columns=columns)
  result_data['TenPct_GEleftX'] = fe_data['raw_X10perc']
  result_data['TenPct_GEleftY'] = fe_data['raw_Y10perc']
  result_data['TenPct_GEleftZ'] = fe_data['raw_Z10perc']
  result_data['TwentyFivePct_GEleftX'] = fe_data['raw_X25perc']
  result_data['TwentyFivePct_GEleftY'] = fe_data['raw_Y25perc']
  result_data['TwentyFivePct_GEleftZ'] = fe_data['raw_Z25perc']
  result_data['FiftyPct_GEleftX'] = fe_data['raw_X50perc']
  result_data['FiftyPct_GEleftY'] = fe_data['raw_Y50perc']
  result_data['FiftyPct_GEleftZ'] = fe_data['raw_Z50perc']
  result_data['SeventyFivePct_GEleftX'] = fe_data['raw_X75perc']
  result_data['SeventyFivePct_GEleftY'] = fe_data['raw_Y75perc']
  result_data['SeventyFivePct_GEleftZ'] = fe_data['raw_Z75perc']
  result_data['NinetyPct_GEleftX'] = fe_data['raw_X90perc']
  result_data['NinetyPct_GEleftY'] = fe_data['raw_Y90perc']
  result_data['NinetyPct_GEleftZ'] = fe_data['raw_Z90perc']

  output_f = f.replace('.csv', '_ANN_input_wrist_2016.txt')
  output_filename = join(ActiPAL_output_folder, output_f)

  result_data.to_csv(output_filename, sep='\t', index=False)
