==> Evaluating_existing_models/Analytics_Scripts/9_combine_results_ANN_thigh_2017.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

from os import listdir
from os.path import isfile, join
import numpy as np
import pandas as pd
import sys, time
import matplotlib.pyplot as plt

epoch_folder = '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/montoye_thigh/Epoch30/'
predictions_folder = '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/montoye_thigh/outputs_from_ANN_thigh_2017/'
output_folder = '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/montoye_thigh/combined_results_ANN_thigh_2017/'

epoch_data_files = [f for f in listdir(epoch_folder) if isfile(join(epoch_folder, f))]
prediction_data_files = [f for f in listdir(predictions_folder) if isfile(join(predictions_folder, f))]
print([len(epoch_data_files),len(prediction_data_files)])

if len(epoch_data_files) == len(prediction_data_files):
    for epoch_file, pred_file in zip(sorted(epoch_data_files), sorted(prediction_data_files)):
      if(epoch_file.split('.csv')[0]==pred_file.split('_predicted_thigh_EE_2017.csv')[0]):
        # print(epoch_file,pred_file)
        result_data = pd.read_csv(epoch_folder + epoch_file)
        predict_data = pd.read_csv(predictions_folder + pred_file)

        result_data['predicted_ee_montoye_2017'] = predict_data['V1']
        # print(result_data.head())
        # print(predict_data['V1'])
        result_data.to_csv((output_folder+pred_file.split('_predicted_thigh_EE_2017.csv')[0]+'_ANN_2017_thigh_combined.csv'), sep=',')
