==> Evaluating_existing_models/Analytics_Scripts/11_get_intensity_minutes_actigraph.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import os
from os import listdir, makedirs
from os.path import join, isfile, isdir, exists
from datetime import datetime
import pandas as pd
import numpy as np
import math
from tqdm import tqdm

# Input
root_folder = '/content/drive/My Drive/OA_Gihan/OA_all/OA activity data_La Trobe'
ActiGraph_60s_epoch_proc_folder = '/content/drive/My Drive/OA_Gihan/OA_all/OA activity data_La Trobe/Analytics_Scripts/Rashmika_Analysis/processed_data/actigraph/epoch_data/60s_ref'
ActiGraph_60hz_raw_folder = '/content/drive/My Drive/OA_Gihan/OA_all/OA activity data_La Trobe/ActiGraph_raw'

# sleep time input_data
sleep_time_file = '/content/drive/My Drive/OA_Gihan/OA_all/OA activity data_La Trobe/Analytics_Scripts/Rashmika_Analysis/metadata/Copy of record_availability_final.csv'

# Output
Actigraph_output_folder = '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/hip_minute_data'

### Load Sleep Data


epoch_files = [f for f in listdir(ActiGraph_60s_epoch_proc_folder) if isfile(join(ActiGraph_60s_epoch_proc_folder, f))]

data_dict = {f.split()[0]: {'epoch': join(ActiGraph_60s_epoch_proc_folder, f)} for f in epoch_files}
print(len(data_dict))

# print(len(epoch_files))
# for ef in epoch_files:
#   key = ef.split('-')[0]
#   if key in data_dict.keys():
#     data_dict[key]['epoch'] = join(ActiPAL_15s_epoch_proc_folder, ef)

print(data_dict['HOA1'])

sleep_df = pd.read_csv(sleep_time_file)
sleep_df.head(2)

### Generate statistical features

# FREQUENCY = 20 #HZ amount of the accelerometer
# REFERENCE_EPOCH = 15 #Epoch value that EE is given
# TIME_EPOCH_DICT = {
#     # 'Epoch1': FREQUENCY * 1,
#     # 'Epoch5': FREQUENCY * 5,
#     'Epoch30': FREQUENCY * 30,
#     # 'Epoch60': FREQUENCY * 60
# }

def met_to_intensity_ground(row):
  ee = complete_df['per_minute_means'][row.name]
  return 1 if ee <= 1.5 else 2 if ee < 3 else 3


def get_bmi_range(row):
  bmi = complete_df['BMI'][row.name]
  if(18.5<=bmi<25):return 'normal'
  elif(25<=bmi<30):return 'overweight'
  elif(30<=bmi):return 'obese'
  else:'error'

s = ['A']*10
print(s)


complete_df = pd.DataFrame()
output_file_name = join(Actigraph_output_folder, 'per_minute_EE_means_hip_reference.csv')
for _, row in tqdm(sleep_df.iterrows(), total=sleep_df.shape[0]):

  user_id = row['Participant']
  sex = row['sex']
  bmi = row['BMI']
 
  begin_time = row['begin_time']
  end_time = row['end_time']
  print(user_id,begin_time,end_time)
  
  temp_df = pd.DataFrame()
  # get respective epoch dataset
  try:
    epoch_df = pd.read_csv(data_dict[user_id]['epoch'])
    epoch_df = epoch_df[(begin_time < epoch_df['Time_stamp']) & (epoch_df['Time_stamp'] < end_time)]
    print("df len",len(epoch_df))
    
    epoch_df_grouped = epoch_df
    per_minute_means = epoch_df_grouped['waist_ee']
    temp_df['per_minute_means'] = per_minute_means
    temp_df['BMI'] = [bmi]*len(per_minute_means)
    temp_df['sex'] = [sex]*len(per_minute_means)
    temp_df.insert(0,'Participant',[user_id]*len(per_minute_means))
    # temp_df['Participant'] = [user_id]*len(hourly_means)
    # prtemp_lenint('',len(temp_df))
    complete_df = pd.concat([complete_df,temp_df])
    # print(len(epoch_df))
    # print(len(epoch_df_grouped))
  except KeyError:
    continue
print('comp',len(complete_df))
complete_df.reset_index(drop=True, inplace=True)
complete_df['Category'] = complete_df.apply(met_to_intensity_ground,axis=1)
complete_df['BMI_range'] = complete_df.apply(get_bmi_range,axis=1)
print("completed")
complete_df.to_csv(output_file_name)



  


complete_df = pd.read_csv('/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/hip_minute_data/per_minute_EE_means_hip_reference.csv')
grouped_df = complete_df.groupby(['Participant','sex','BMI_range', 'Category']).size().unstack(fill_value=0)
output_g_file_name = join(Actigraph_output_folder, 'per_minute_EE_means_hip_reference_grouped.csv')
grouped_df.to_csv(output_g_file_name)


# for _, row in tqdm(complete_df.iterrows(), total=complete_df.shape[0]):
#   print(row)

all_df = pd.read_csv('/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/hip_minute_data/per_minute_EE_means_hip_reference_grouped.csv')

female = all_df['sex'] == 'female'
male = all_df['sex'] == 'male'
obese = all_df['BMI_range'] == 'obese'
overweight = all_df['BMI_range'] == 'overweight'
normal = all_df['BMI_range'] == 'normal'


male_df = all_df[male]
female_df = all_df[female]
obese_df = all_df[obese]
overweight_df = all_df[overweight]
normal_df = all_df[normal]

# print(['all,male,female',len(all_df),len(male_df),len(female_df)])
# print(['all,male,female',all_df['1'].mean,len(male_df),len(female_df)])

with pd.ExcelWriter(join(Actigraph_output_folder,'hip_reference_minute_values.xlsx')) as writer: 
    all_df.to_excel(writer, sheet_name='all')
    female_df.to_excel(writer, sheet_name='female')
    male_df.to_excel(writer, sheet_name='male')
    obese_df.to_excel(writer, sheet_name='obese')
    overweight_df.to_excel(writer, sheet_name='overweight')
    normal_df.to_excel(writer, sheet_name='normal')

grouped_df = pd.read_csv('/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/hip_minute_data/per_minute_EE_means_hip_grouped_participant.csv')

female = grouped_df['sex'] == 'female'
male = grouped_df['sex'] == 'male'
grouped_df[female].to_csv(join(Actigraph_output_folder,'female_categorized.csv'))
grouped_df[male].to_csv(join(Actigraph_output_folder,'male_categorized.csv'))

grouped_df = pd.read_csv('/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/hip_minute_data/per_minute_EE_means_hip_grouped_participant.csv')

obese = grouped_df['BMI_range'] == 'obese'
overweight = grouped_df['BMI_range'] == 'overweight'
normal = grouped_df['BMI_range'] == 'normal'
grouped_df[obese].to_csv(join(Actigraph_output_folder,'obese_categorized.csv'))
grouped_df[overweight].to_csv(join(Actigraph_output_folder,'overweight_categorized.csv'))
grouped_df[normal].to_csv(join(Actigraph_output_folder,'normal_categorized.csv'))

data = pd.DataFrame()
data['X'] = [1,2,3,4,5,6,7,8,9]
data['Y'] = [0,2,4,6,8,10,12,14,16]
data['Z'] = [3,6,9,12,15,18,21,24,27]

grouped = data.groupby(data.index // 2)
new_d = data.iloc[::2]
new_d['X_sum'] = grouped['X'].sum()
print(data)
# print(grouped)
print(new_d)
# print(mean_g)

complete_df = pd.DataFrame()
complete_df['Participant'] = [1,1,3,4,5,6,7,8,9,10]
complete_df['per_minute_means']=[1.2,3,4,1.1,0.4,4,1,2,3,6]
complete_df['Category'] = complete_df.apply(met_to_intensity_ground,axis=1)
complete_df['sex'] = ['male','female','female','male','male','female','female','male','male','female']
complete_df['BMI'] = [22,26,19,22,31,28,30,26,32,34]
complete_df['BMI_range'] = complete_df.apply(get_bmi_range,axis=1)
grouped_df = complete_df.groupby(['BMI_range', 'Category']).size().unstack(fill_value=0)
print(complete_df)
print(grouped_df)
