==> Evaluating_existing_models/Analytics_Scripts/12_get_intensity_minutes_summary_thigh.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import os
from os import listdir, makedirs
from os.path import join, isfile, isdir, exists
from datetime import datetime
import pandas as pd
import numpy as np
import math
from tqdm import tqdm

import csv
excel_folder = '/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/thigh_minute_data/'

def get_res_cat(name,df):
  results = []
  for k in ['1','2','3']:
    try:
      value = (str(round(df[k].mean(),3))+' ± '+str(round(df[k].std(),3))).strip()
      results_temp = [name+'_'+k,len(df[k]),round(df[k].mean(),3),round(df[k].std(),3),value]
      results.append(results_temp)
    except Exception:
      continue
  return results

def get_results_for_book(book_name,model_name):
  all = pd.read_excel(open(excel_folder+book_name, 'rb'),sheet_name='all')
  female = pd.read_excel(open(excel_folder+book_name, 'rb'),sheet_name='female')
  male = pd.read_excel(open(excel_folder+book_name, 'rb'),sheet_name='male')
  obese = pd.read_excel(open(excel_folder+book_name, 'rb'),sheet_name='obese')
  overweight = pd.read_excel(open(excel_folder+book_name, 'rb'),sheet_name='overweight')
  normal = pd.read_excel(open(excel_folder+book_name, 'rb'),sheet_name='normal')
  results = []
  results.append(get_res_cat(model_name+'_all',all))
  results.append(get_res_cat(model_name+'_female',female))
  results.append(get_res_cat(model_name+'_male',male))
  results.append(get_res_cat(model_name+'_obese',obese))
  results.append(get_res_cat(model_name+'_overweight',overweight))
  results.append(get_res_cat(model_name+'_normal',normal))

  return results

all_results = []
all_results.append(get_results_for_book('1_thigh_reference_minute_values.xlsx','reference_thigh'))
all_results.append(get_results_for_book('4_thigh_mont_1_minute_values.xlsx','mont_1_thigh'))
all_results.append(get_results_for_book('4_thigh_mont_2_minute_values.xlsx','mont_2_thigh'))
all_results.append(get_results_for_book('4_thigh_mont_3_minute_values.xlsx','mont_3_thigh'))
all_results.append(get_results_for_book('4_thigh_mont_4_minute_values.xlsx','mont_4_thigh'))
all_results.append(get_results_for_book('4_thigh_mont_17_minute_values.xlsx','mont_17_thigh'))

all_results.append(get_results_for_book('4_thigh_white_1_minute_values.xlsx','white_1_thigh'))
all_results.append(get_results_for_book('4_thigh_white_2_minute_values.xlsx','white_2_thigh'))


results_file = excel_folder+'5_summary_thigh.csv'
f = open(results_file, 'w', encoding='utf-8')
writer = csv.writer(f)
writer.writerow(['model','count','mean','standard_deviation','value'])
for l in all_results:
  for ele in l:
    for el in ele:
      print(el)
      writer.writerow(el)

f.close()

### Generate statistical features


def met_to_intensity_w1(row):
  ee = per_person_df['white_1'][row.name]
  return 1 if ee <= 1.5 else 2 if ee < 3 else 3

def met_to_intensity_w2(row):
  ee = per_person_df['white_2'][row.name]
  return 1 if ee <= 1.5 else 2 if ee < 3 else 3


def get_bmi_range(row):
  bmi = per_person_df['BMI'][row.name]
  if(18.5<=bmi<25):return 'normal'
  elif(25<=bmi<30):return 'overweight'
  elif(30<=bmi):return 'obese'
  else:'error'

s = ['A']*10
print(s)

info_dict = {}

sleep_time_file = '/content/drive/My Drive/OA_Gihan/OA_all/OA activity data_La Trobe/Analytics_Scripts/Rashmika_Analysis/metadata/Copy of record_availability_final.csv'
sleep_df = pd.read_csv(sleep_time_file)
BMIs = []
# sleep_df.head(2)
for _, row in tqdm(sleep_df.iterrows(), total=sleep_df.shape[0]):
  user_id = row['Participant']
  sex = row['sex']
  BMI = row['BMI']
  
  if(sex!='male' and sex!='female'):
    if(user_id in combined_ids):
      print(user_id)
      sex='female'
  if(str(BMI)=='nan'):
    BMI = sum(BMIs)/len(BMIs)
  else:
    BMIs.append(BMI)
  info_dict[user_id] = [sex,BMI]

print(info_dict)
print(sum(BMIs)/len(BMIs))




all_df = pd.DataFrame()
for each_id in combined_ids:
    sex = info_dict[each_id][0]
    bmi = info_dict[each_id][1]
    complete_df = pd.DataFrame()
    # print(combined_data_files)
    for combined_file in combined_data_files:
        
        if (combined_file.split('_')[0] != each_id): continue
        print(combined_file)
        combined_data = pd.read_csv(combined_results_folder + combined_file)
        print(len(combined_data))
        # sex = combined_data['sex'][0]
        # bmi = combined_data['BMI'][0]
        
        temp_df = combined_data[['EE_white_1',
       'EE_white_2']]
        
        complete_df = pd.concat([complete_df, temp_df])
        
    

    complete_df_grouped = complete_df.groupby(complete_df.index // 12)
    per_person_df = pd.DataFrame()
    per_person_df['Participant'] = [each_id]*len(complete_df_grouped)
    per_person_df['sex'] = [sex]*len(complete_df_grouped)
    per_person_df['BMI'] = [bmi]*len(complete_df_grouped)
    per_person_df['white_1'] = complete_df_grouped['EE_white_1'].mean()
    per_person_df['white_2'] = complete_df_grouped['EE_white_2'].mean()
    per_person_df['white_1_cat'] = per_person_df.apply(met_to_intensity_w1,axis=1)
    per_person_df['white_2_cat'] = per_person_df.apply(met_to_intensity_w2,axis=1)
    per_person_df['BMI_range'] = per_person_df.apply(get_bmi_range,axis=1)
    all_df = pd.concat([all_df, per_person_df])
    

# print(len(all_df))
out_file = output_folder+'2_per_minute_EE_means_thigh_white.csv'
print(out_file)
all_df.to_csv(out_file)

  


complete_df = pd.read_csv('/content/drive/My Drive/OA activity data_La Trobe/Analytics_Scripts/Gihan_analysis/processed data/thigh_minute_data/2_per_minute_EE_means_thigh_white.csv')


grouped_df_m1 = complete_df.groupby(['Participant','sex','BMI_range','white_1_cat']).size().unstack(fill_value=0)
grouped_df_m1.to_csv(join(ActiPAL_output_folder, '3_per_minute_EE_means_thigh_white_1_grouped.csv'))

grouped_df_m2 = complete_df.groupby(['Participant','sex','BMI_range','white_2_cat']).size().unstack(fill_value=0)
grouped_df_m2.to_csv(join(ActiPAL_output_folder, '3_per_minute_EE_means_thigh_white_2_grouped.csv'))


def categorize(input_name,output_sheet_name):
  all_df = pd.read_csv(join(ActiPAL_output_folder,input_name))

  female = all_df['sex'] == 'female'
  male = all_df['sex'] == 'male'
  obese = all_df['BMI_range'] == 'obese'
  overweight = all_df['BMI_range'] == 'overweight'
  normal = all_df['BMI_range'] == 'normal'


  male_df = all_df[male]
  female_df = all_df[female]
  obese_df = all_df[obese]
  overweight_df = all_df[overweight]
  normal_df = all_df[normal]

  # print(['all,male,female',len(all_df),len(male_df),len(female_df)])
  # print(['all,male,female',all_df['1'].mean,len(male_df),len(female_df)])

  with pd.ExcelWriter(join(ActiPAL_output_folder,output_sheet_name)) as writer: 
      all_df.to_excel(writer, sheet_name='all')
      female_df.to_excel(writer, sheet_name='female')
      male_df.to_excel(writer, sheet_name='male')
      obese_df.to_excel(writer, sheet_name='obese')
      overweight_df.to_excel(writer, sheet_name='overweight')
      normal_df.to_excel(writer, sheet_name='normal')


categorize('3_per_minute_EE_means_thigh_white_1_grouped.csv','4_thigh_white_1_minute_values.xlsx')
categorize('3_per_minute_EE_means_thigh_white_2_grouped.csv','4_thigh_white_2_minute_values.xlsx')



data = pd.DataFrame()
data['X'] = [1,2,3,4,5,6,7,8,9]
data['Y'] = [0,2,4,6,8,10,12,14,16]
data['Z'] = [3,6,9,12,15,18,21,24,27]

grouped = data.groupby(data.index // 2)
new_d = data.iloc[::2]
new_d['X_sum'] = grouped['X'].sum()
print(data)
# print(grouped)
print(new_d)
# print(mean_g)
